// src/app/models/station.ts
export interface Station {
    id?: number;
    Nom_Station: String;
    Date_description: String;
    Secteur: String;
    Quartier: String;
    Localistaion: String;
    vocation: String;
    Phase_physiologique_dominante: String;
    Etat_general: String;
    Conduite: String;
    Revetement: String;
    Nature_substrat: String;
    protection: String;
    Irrigation: String;
    Preconisation: String;
    Taille: String;
    Arrondissement: String;
    Zone: String;
  
}